+++ src/containers/auth/PasswordScreen.js
@@ -3,13 +3,8 @@ import PropTypes from 'prop-types';
 import { inject, observer } from 'mobx-react';
 import Password from '../../components/auth/Password';
 import UserStore from '../../stores/UserStore';
-import { gaPage } from '../../lib/analytics';
 
 export default @inject('stores', 'actions') @observer class PasswordScreen extends Component {
-  componentDidMount() {
-    gaPage('Auth/Password Retrieve');
-  }
-
   render() {
     const { actions, stores } = this.props;