+++ src/containers/auth/WelcomeScreen.js
@@ -5,13 +5,8 @@ import { inject, observer } from 'mobx-react';
 import Welcome from '../../components/auth/Welcome';
 import UserStore from '../../stores/UserStore';
 import RecipePreviewsStore from '../../stores/RecipePreviewsStore';
-import { gaPage } from '../../lib/analytics';
 
 export default @inject('stores', 'actions') @observer class LoginScreen extends Component {
-  componentDidMount() {
-    gaPage('Auth/Welcome');
-  }
-
   render() {
     const { user, recipePreviews } = this.props.stores;