+++ src/lib/Menu.js
@@ -1,5 +1,5 @@
 import { remote, shell } from 'electron';
-import { observable, autorun, computed } from 'mobx';
+import { observable, autorun } from 'mobx';
 import { defineMessages } from 'react-intl';
 
 import { isMac, ctrlKey, cmdKey } from '../environment';
@@ -179,6 +179,22 @@ const menuItems = defineMessages({
     id: 'menu.services.addNewService',
     defaultMessage: '!!!Add New Service...',
   },
+  activateNextService: {
+    id: 'menu.services.setNextServiceActive',
+    defaultMessage: '!!!Activate next service...',
+  },
+  activatePreviousService: {
+    id: 'menu.services.activatePreviousService',
+    defaultMessage: '!!!Activate previous service...',
+  },
+  muteApp: {
+    id: 'sidebar.muteApp',
+    defaultMessage: '!!!Disable notifications & audio',
+  },
+  unmuteApp: {
+    id: 'sidebar.unmuteApp',
+    defaultMessage: '!!!Enable notifications & audio',
+  },
 });
 
 function getActiveWebview() {
@@ -239,16 +255,32 @@ const _templateFactory = intl => [
       },
       {
         label: intl.formatMessage(menuItems.resetZoom),
-        role: 'resetzoom',
+        accelerator: 'Cmd+0',
+        click() {
+          getActiveWebview().setZoomLevel(0);
+        },
       },
       {
         label: intl.formatMessage(menuItems.zoomIn),
-        // accelerator: 'Cmd+=',
-        role: 'zoomin',
+        accelerator: 'Cmd+plus',
+        click() {
+          const activeService = getActiveWebview();
+          activeService.getZoomLevel((level) => {
+            // level 9 =~ +300% and setZoomLevel wouldnt zoom in further
+            if (level < 9) activeService.setZoomLevel(level + 1);
+          });
+        },
       },
       {
         label: intl.formatMessage(menuItems.zoomOut),
-        role: 'zoomout',
+        accelerator: 'Cmd+-',
+        click() {
+          const activeService = getActiveWebview();
+          activeService.getZoomLevel((level) => {
+            // level -9 =~ -50% and setZoomLevel wouldnt zoom out further
+            if (level > -9) activeService.setZoomLevel(level - 1);
+          });
+        },
       },
       {
         type: 'separator',
@@ -392,10 +424,12 @@ const _titleBarTemplateFactory = intl => [
       },
       {
         label: intl.formatMessage(menuItems.zoomIn),
-        accelerator: `${ctrlKey}+Plus`,
+        accelerator: `${ctrlKey}+=`,
         click() {
-          getActiveWebview().getZoomLevel((zoomLevel) => {
-            getActiveWebview().setZoomLevel(zoomLevel === 5 ? zoomLevel : zoomLevel + 1);
+          const activeService = getActiveWebview();
+          activeService.getZoomLevel((level) => {
+            // level 9 =~ +300% and setZoomLevel wouldnt zoom in further
+            if (level < 9) activeService.setZoomLevel(level + 1);
           });
         },
       },
@@ -403,8 +437,10 @@ const _titleBarTemplateFactory = intl => [
         label: intl.formatMessage(menuItems.zoomOut),
         accelerator: `${ctrlKey}+-`,
         click() {
-          getActiveWebview().getZoomLevel((zoomLevel) => {
-            getActiveWebview().setZoomLevel(zoomLevel === -5 ? zoomLevel : zoomLevel - 1);
+          const activeService = getActiveWebview();
+          activeService.getZoomLevel((level) => {
+            // level -9 =~ -50% and setZoomLevel wouldnt zoom out further
+            if (level > -9) activeService.setZoomLevel(level - 1);
           });
         },
       },
@@ -499,13 +535,14 @@ export default class FranzMenu {
   }
 
   _build() {
-    const serviceTpl = Object.assign([], this.serviceTpl); // need to clone object so we don't modify computed (cached) object
+    // need to clone object so we don't modify computed (cached) object
+    const serviceTpl = Object.assign([], this.serviceTpl());
 
     if (window.franz === undefined) {
       return;
     }
 
-    const intl = window.franz.intl;
+    const { intl } = window.franz;
     const tpl = isMac ? _templateFactory(intl) : _titleBarTemplateFactory(intl);
 
     tpl[1].submenu.push({
@@ -663,17 +700,6 @@ export default class FranzMenu {
       }, about);
     }
 
-    serviceTpl.unshift({
-      label: intl.formatMessage(menuItems.addNewService),
-      accelerator: `${cmdKey}+N`,
-      click: () => {
-        this.actions.ui.openSettings({ path: 'recipes' });
-      },
-      enabled: this.stores.user.isLoggedIn,
-    }, {
-      type: 'separator',
-    });
-
     if (serviceTpl.length > 0) {
       tpl[3].submenu = serviceTpl;
     }
@@ -683,22 +709,49 @@ export default class FranzMenu {
     Menu.setApplicationMenu(menu);
   }
 
-  @computed get serviceTpl() {
-    const services = this.stores.services.allDisplayed;
+  serviceTpl() {
+    const { intl } = window.franz;
+    const { user, services, settings } = this.stores;
+    if (!user.isLoggedIn) return [];
+    const menu = [];
 
-    if (this.stores.user.isLoggedIn) {
-      return services.map((service, i) => ({
-        label: this._getServiceName(service),
-        accelerator: i < 9 ? `${cmdKey}+${i + 1}` : null,
-        type: 'radio',
-        checked: service.isActive,
-        click: () => {
-          this.actions.service.setActive({ serviceId: service.id });
-        },
-      }));
-    }
+    menu.push({
+      label: intl.formatMessage(menuItems.addNewService),
+      accelerator: `${cmdKey}+N`,
+      click: () => {
+        this.actions.ui.openSettings({ path: 'recipes' });
+      },
+    }, {
+      type: 'separator',
+    }, {
+      label: intl.formatMessage(menuItems.activateNextService),
+      accelerator: `${cmdKey}+alt+right`,
+      click: () => this.actions.service.setActiveNext(),
+    }, {
+      label: intl.formatMessage(menuItems.activatePreviousService),
+      accelerator: `${cmdKey}+alt+left`,
+      click: () => this.actions.service.setActivePrev(),
+    }, {
+      label: intl.formatMessage(
+        settings.all.app.isAppMuted ? menuItems.unmuteApp : menuItems.muteApp,
+      ).replace('&', '&&'),
+      accelerator: `${cmdKey}+shift+m`,
+      click: () => this.actions.app.toggleMuteApp(),
+    }, {
+      type: 'separator',
+    });
 
-    return [];
+    services.allDisplayed.forEach((service, i) => (menu.push({
+      label: this._getServiceName(service),
+      accelerator: i < 9 ? `${cmdKey}+${i + 1}` : null,
+      type: 'radio',
+      checked: service.isActive,
+      click: () => {
+        this.actions.service.setActive({ serviceId: service.id });
+      },
+    })));
+
+    return menu;
   }
 
   _getServiceName(service) {