+++ src/stores/AppStore.js
@@ -1,21 +1,25 @@
 import { remote, ipcRenderer, shell } from 'electron';
-import { action, computed, observable } from 'mobx';
+import {
+  action, computed, observable, reaction,
+} from 'mobx';
 import moment from 'moment';
-import key from 'keymaster';
 import { getDoNotDisturb } from '@meetfranz/electron-notification-state';
 import AutoLaunch from 'auto-launch';
 import prettyBytes from 'pretty-bytes';
+import ms from 'ms';
+import { URL } from 'url';
 
 import Store from './lib/Store';
 import Request from './lib/Request';
 import { CHECK_INTERVAL, DEFAULT_APP_SETTINGS } from '../config';
-import { isMac, isLinux, isWindows } from '../environment';
+import { isMac } from '../environment';
 import locales from '../i18n/translations';
-import { gaEvent } from '../lib/analytics';
+import { gaEvent, gaPage } from '../lib/analytics';
 import { onVisibilityChange } from '../helpers/visibility-helper';
 import { getLocale } from '../helpers/i18n-helpers';
 
 import { getServiceIdsFromPartitions, removeServicePartitionDirectory } from '../helpers/service-helpers.js';
+import { isValidExternalURL } from '../helpers/url-helpers';
 
 const debug = require('debug')('Franz:AppStore');
 
@@ -110,12 +114,12 @@ export default class AppStore extends Store {
     // Check if system is muted
     // There are no events to subscribe so we need to poll everey 5s
     this._systemDND();
-    setInterval(() => this._systemDND(), 5000);
+    setInterval(() => this._systemDND(), ms('5s'));
 
     // Check for updates once every 4 hours
     setInterval(() => this._checkForUpdates(), CHECK_INTERVAL);
     // Check for an update in 30s (need a delay to prevent Squirrel Installer lock file issues)
-    setTimeout(() => this._checkForUpdates(), 30000);
+    setTimeout(() => this._checkForUpdates(), ms('30s'));
     ipcRenderer.on('autoUpdate', (event, data) => {
       if (data.available) {
         this.updateStatus = this.updateStatusTypes.AVAILABLE;
@@ -152,27 +156,6 @@ export default class AppStore extends Store {
       this.stores.router.push(url);
     });
 
-    // Set active the next service
-    key(
-      '⌘+pagedown, ctrl+pagedown, ⌘+alt+right, ctrl+tab', () => {
-        this.actions.service.setActiveNext();
-      },
-    );
-
-    // Set active the prev service
-    key(
-      '⌘+pageup, ctrl+pageup, ⌘+alt+left, ctrl+shift+tab', () => {
-        this.actions.service.setActivePrev();
-      },
-    );
-
-    // Global Mute
-    key(
-      '⌘+shift+m ctrl+shift+m', () => {
-        this.actions.app.toggleMuteApp();
-      },
-    );
-
     this.locale = this._getDefaultLocale();
 
     this._healthCheck();
@@ -184,6 +167,12 @@ export default class AppStore extends Store {
 
       debug('Window is visible/focused', isVisible);
     });
+
+    // analytics autorun
+    reaction(() => this.stores.router.location.pathname, (pathname) => {
+      gaPage(pathname);
+    });
+    console.log('router location', this.stores.router.location);
   }
 
   @computed get cacheSize() {
@@ -196,7 +185,15 @@ export default class AppStore extends Store {
   }) {
     if (this.stores.settings.all.app.isAppMuted) return;
 
+    // TODO: is there a simple way to use blobs for notifications without storing them on disk?
+    if (options.icon.startsWith('blob:')) {
+      delete options.icon;
+    }
+
     const notification = new window.Notification(title, options);
+
+    debug('New notification', title, options);
+
     notification.onclick = (e) => {
       if (serviceId) {
         this.actions.service.sendIPCMessage({
@@ -206,12 +203,13 @@ export default class AppStore extends Store {
         });
 
         this.actions.service.setActive({ serviceId });
-
-        if (isWindows) {
+        mainWindow.show();
+        if (app.mainWindow.isMinimized()) {
           mainWindow.restore();
-        } else if (isLinux) {
-          mainWindow.show();
         }
+        mainWindow.focus();
+
+        debug('Notification click handler');
       }
     };
   }
@@ -247,7 +245,14 @@ export default class AppStore extends Store {
   }
 
   @action _openExternalUrl({ url }) {
-    shell.openExternal(url);
+    const parsedUrl = new URL(url);
+    debug('open external url', parsedUrl);
+
+    if (isValidExternalURL(url)) {
+      shell.openExternal(url);
+    }
+
+    gaEvent('External URL', 'open', parsedUrl.host);
   }
 
   @action _checkForUpdates() {
@@ -271,7 +276,6 @@ export default class AppStore extends Store {
 
   @action _muteApp({ isMuted, overrideSystemMute = true }) {
     this.isSystemMuteOverridden = overrideSystemMute;
-
     this.actions.settings.update({
       type: 'app',
       data: {
@@ -308,7 +312,7 @@ export default class AppStore extends Store {
     } else {
       const deltaTime = moment().diff(this.timeOfflineStart);
 
-      if (deltaTime > 30 * 60 * 1000) {
+      if (deltaTime > ms('30m')) {
         this.actions.service.reloadAll();
       }
     }