+++ src/stores/FeaturesStore.js
@@ -7,6 +7,7 @@ import delayApp from '../features/delayApp';
 import spellchecker from '../features/spellchecker';
 import serviceProxy from '../features/serviceProxy';
 import basicAuth from '../features/basicAuth';
+import shareFranz from '../features/shareFranz';
 
 import { DEFAULT_FEATURES_CONFIG } from '../config';
 
@@ -56,5 +57,6 @@ export default class FeaturesStore extends Store {
     spellchecker(this.stores, this.actions);
     serviceProxy(this.stores, this.actions);
     basicAuth(this.stores, this.actions);
+    shareFranz(this.stores, this.actions);
   }
 }